+++ bot/src/main/java/edu/java/bot/configuration/ApplicationConfig.java
@@ -3,6 +3,7 @@ package edu.java.bot.configuration;
 import jakarta.validation.constraints.NotEmpty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
+import utils.retry.model.RetryConfig;
 
 @Validated
 @ConfigurationProperties(prefix = "app", ignoreUnknownFields = false)
@@ -10,6 +11,8 @@ public record ApplicationConfig(
     @NotEmpty
     String telegramToken,
     @NotEmpty
-    String scrapperApiUrl
+    String scrapperApiUrl,
+    RetryConfig retry
 ) {
+
 }
