+++ bot/src/main/java/edu/java/bot/configuration/ScrapperClientConfig.java
@@ -3,6 +3,7 @@ package edu.java.bot.configuration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.reactive.function.client.WebClient;
+import utils.retry.RetryUtils;
 
 @Configuration
 public class ScrapperClientConfig {
@@ -16,6 +17,7 @@ public class ScrapperClientConfig {
     public WebClient scrapperWebClient() {
         return WebClient
             .builder()
+            .filter(RetryUtils.getRetryFilter("scrapper", applicationConfig.retry()))
             .baseUrl(applicationConfig.scrapperApiUrl())
             .build();
     }
