+++ bot/src/main/resources/application.yml
@@ -1,6 +1,16 @@
 app:
   telegram-token: ${TELEGRAM_BOT_TOKEN}
-  scrapper_api_url: ${SCRAPPER_API_URL:http://localhost:8080}
+  scrapper_api_url: ${SCRAPPER_API_URL:http://localhost:8080}\
+
+  retry:
+    retryItems:
+      - client-name: scrapper
+        type: linear
+        max-attempts: 5
+        min-backoff: 1s
+        max-backoff: 15s
+        increment: 3s
+        error-codes: 500
 
 spring:
   application:
