+++ common/src/main/java/utils/retry/model/RetryConfig.java
@@ -1,0 +1,21 @@
+package utils.retry.model;
+
+import jakarta.validation.constraints.NotNull;
+import java.time.Duration;
+import java.util.List;
+
+public record RetryConfig(
+    List<RetryItem> retryItems
+) {
+    public record RetryItem(
+        @NotNull String clientName,
+        @NotNull RetryType type,
+        Long maxAttempts,
+        Duration minBackoff,
+        Duration maxBackoff,
+        Double jitter,
+        Duration increment,
+        List<Integer> errorCodes
+    ) {
+    }
+}
