+++ common/src/main/java/utils/retry/model/RetryType.java
@@ -1,0 +1,7 @@
+package utils.retry.model;
+
+public enum RetryType {
+    CONSTANT,
+    LINEAR,
+    EXPONENTIAL
+}
