+++ scrapper/src/main/java/edu/java/configuration/ApplicationConfig.java
@@ -2,10 +2,10 @@ package edu.java.configuration;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.Duration;
-import java.util.List;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.validation.annotation.Validated;
+import utils.retry.model.RetryConfig;
 
 @Validated
 @ConfigurationProperties(prefix = "app", ignoreUnknownFields = false)
@@ -28,25 +28,4 @@ public record ApplicationConfig(
         JOOQ
     }
 
-    public record RetryConfig(
-        List<RetryItem> retryItems
-    ) {
-        public record RetryItem(
-            @NotNull String clientName,
-            @NotNull RetryType type,
-            int maxAttempts,
-            Duration minBackoff,
-            Duration maxBackoff,
-            double jitter,
-            Duration increment,
-            List<Integer> errorCodes
-        ) {
-        }
-
-        public enum RetryType {
-            CONSTANT,
-            LINEAR,
-            EXPONENTIAL
-        }
-    }
 }
